refactor redux/store to use different slices
