+++ src/redux/store.ts
@@ -1,0 +1,14 @@
+// store.js
+import { configureStore } from '@reduxjs/toolkit'
+import { settingsReducer } from './slices/settings'
+import { gameReducer } from './slices/game'
+
+const store = configureStore({
+  reducer: {
+    settings: settingsReducer,
+    game: gameReducer
+  }
+})
+
+export type RootState = ReturnType<typeof store.getState>
+export default store
